# docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "JEC Appointment Go Service",
        "title": "JEC Appointment Go Service",
        "contact": {
            "name": "JEC",
            "email": "no-reply@jec.co.id"
        },
        "version": "1.0.0"
    },
    "host": "jec.co.id",
    "basePath": "/",
    "paths": {
        "/appointment": {
            "get": {
                "description": "Get Appointment List.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Appointment"
                ],
                "summary": "List of Appointment Object.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Healthcare ID",
                        "name": "healthcare_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Appointment No",
                        "name": "appointment_no",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Patient ID",
                        "name": "patient_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Paramedic No",
                        "name": "paramedic_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "no found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "description": "Update Appointment.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Appointment"
                ],
                "summary": "A newly updated Appointment.",
                "parameters": [
                    {
                        "description": "Data request",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain_appointment.AppointmentUpdateProto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "201": {
                        "description": "created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "204": {
                        "description": "no content",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "no found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "Create Appointment.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Appointment"
                ],
                "summary": "A newly created Appointment.",
                "parameters": [
                    {
                        "description": "Data request",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain_appointment.AppointmentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "204": {
                        "description": "no content",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete Appointment.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Appointment"
                ],
                "summary": "A newly deleted Appointment.",
                "parameters": [
                    {
                        "description": "Data request",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain_appointment.AppointmentDeleteProto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "202": {
                        "description": "accepted",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "no found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/appointment/get": {
            "get": {
                "description": "Get Appointment By Healthcare ID and Appointment No.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Appointment"
                ],
                "summary": "Appointment Object.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Healthcare ID",
                        "name": "healthcare_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Appointment No",
                        "name": "appointment_no",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "no found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain_appointment.AppointmentDeleteProto": {
            "type": "object",
            "properties": {
                "deleteappointment": {
                    "$ref": "#/definitions/domain_appointment.AppointmentProto"
                }
            }
        },
        "domain_appointment.AppointmentProto": {
            "type": "object",
            "properties": {
                "appointment_date": {
                    "$ref": "#/definitions/timestamppb.Timestamp"
                },
                "appointment_no": {
                    "type": "string"
                },
                "appointment_time": {
                    "$ref": "#/definitions/timestamppb.Timestamp"
                },
                "create_at": {
                    "$ref": "#/definitions/timestamppb.Timestamp"
                },
                "healthcare_id": {
                    "type": "string"
                },
                "is_void": {
                    "type": "boolean"
                },
                "paramedic_id": {
                    "type": "string"
                },
                "patient_id": {
                    "type": "string"
                },
                "schedule_slot_id": {
                    "type": "integer"
                },
                "user_create": {
                    "type": "string"
                }
            }
        },
        "domain_appointment.AppointmentRequest": {
            "type": "object",
            "properties": {
                "AppointmentDate": {
                    "type": "string"
                },
                "AppointmentNo": {
                    "type": "string"
                },
                "AppointmentTime": {
                    "type": "string"
                },
                "CreateAt": {
                    "type": "string"
                },
                "HealthcareId": {
                    "type": "string"
                },
                "IsVoid": {
                    "type": "boolean"
                },
                "ParamedicId": {
                    "type": "string"
                },
                "PatientId": {
                    "type": "string"
                },
                "ScheduleSlotId": {
                    "type": "integer"
                },
                "UserCreate": {
                    "type": "string"
                }
            }
        },
        "domain_appointment.AppointmentUpdateProto": {
            "type": "object",
            "properties": {
                "updateappointment": {
                    "$ref": "#/definitions/domain_appointment.AppointmentProto"
                }
            }
        },
        "timestamppb.Timestamp": {
            "type": "object",
            "properties": {
                "nanos": {
                    "description": "Non-negative fractions of a second at nanosecond resolution. Negative\nsecond values with fractions must still have non-negative nanos values\nthat count forward in time. Must be from 0 to 999,999,999\ninclusive.",
                    "type": "integer"
                },
                "seconds": {
                    "description": "Represents seconds of UTC time since Unix epoch\n1970-01-01T00:00:00Z. Must be from 0001-01-01T00:00:00Z to\n9999-12-31T23:59:59Z inclusive.",
                    "type": "integer"
                }
            }
        }
    }
}